refactor(sdk): split relaycast.ts into translate/placement/client modules
Decompose the 1654-line RelaycastMessagingClient file into single-
responsibility siblings behind an unchanged public surface:

- relaycast-translate.ts: pure wire<->domain translation and record readers
- relaycast-placement.ts: RelayPlacementError, placement payload helpers
- relaycast-client.ts: structural relaycast adapter types, options, factory

relaycast.ts (now 1126 lines) keeps the stateful client and re-exports
RelayPlacementError and RelaycastMessagingOptions so consumers are unaffected.
Pure move: no behavior change. Verified with build, full test suite (959
pass), and an SDK public-surface smoke test.
